Add files via upload
These files are technical references in the field of adaptive materials design.  Since this is a branch of global optimization theory, it may be useful to see the similarities with respect to reinforcement learning. Both use explore/exploit methodologies, but the materials design avenue uses uncertainty and knowledge gradients to choose.